pyproject.toml
[project]
name = "wri-ai-experiments"
version = "0.1.0"
description = "Monorepo for WRI Applied AI Group experiments"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }

# Core dependencies for repo management scripts
dependencies = [
    "marimo>=0.19.7",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "ruff>=0.14.0",
    "pytest>=8.0.0",
]

# For notebook experiments
notebooks = [
    "jupyter>=1.0.0",
    "nbconvert>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["packages"]

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["wri"]

# pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests", "packages/*/tests", "experiments/*/tests"]
python_files = ["test_*.py", "*_test.py"]
